add tiny in its own component
